+++ CHANGELOG.md
@@ -1,5 +1,27 @@
 # CHANGELOG
 
+## V0.3.1 | 2021-12-08
+### Fixed
+- Code small
+- HTTP: Deep merge request parameters and configs
+
+## V0.3.0 | 2021-07-20
+### Features
+- HTTP: API supports path parameters in URL, e.g. /user/{id}
+- HTTP: Interceptors of request, response and error
+- HTTP: Mock is disabled when the enviroment MOCK is "none"
+
+### Changed
+- HTTP: HttpAPI extends HttpRequest
+- HTTP: Mock skips only request and not response/error
+
+### Fixed
+- Model: Code small
+
+### Removed
+- Model: React useModel
+- HTTP: The enviroment MOCK_DISABLED
+
 ## V0.2.1
 ### Features
 - Storage: *remove*